Add tests
I added a series of tests to determine if the outcome of the model was
the same with or without events. I was curious since you updated the
core solver in 5d30fa2. I also ran these tested on afe6be5 and they
had the same results. Curious since I assumed that the two solvers
would behave differently in this respect. Perhaps I am missing
something about the way my events are being processed by the solver
and therefore I may not be testing what I think I am testing.

The goal is to produce a test that shows that the scheduled events do
not change the rate of the stochastic events.

TEST1 PASSES: Model with selection from 1 compartment and shifting nowhere

TEST2 PASSES: A model where an individual is shifted in both
directions on each day.

TEST3 FAILS: This is a model with selection of individuals from both
compartments and shifting them nowhere

UPDATE

Now I understand that TEST3 fails because the individuals are sampled
from 2 compartments and this requires the use of a random number. When
this random number is consumed it changes the subsequent stream and
therefore the stochastic event change.

TESTS 1 and 2 are not doing what I thought because the time to next
event that is calculated to be further in the future than one time
step is alway discarded anyway since the simulator stops every
tiemstep and checks if it needs to execute any scheduled
events. Therefore the with inert events and without is identical.

However, I now also better understand the logic behind discarding
these values of time-to-event that is greater than one tiem unit:
One must remember that the time-to-event is sampled from a
distribution and that more rarely a small value wil lbe selected and
that then the cumulative sum of the samples both discarded and used
become the time to the next event and that this is equivalent sampling once.