during initialization phase, calculate variance from last particle
Since we are initializing the chain we can just calculcate the
variance by the 'rule of thumb' value of (param/10)^2 where param is
the last accepted parameter (param = object@chain[(i - 1), j]). This
will allow som drift in the parameters during the initialization
phase. The alternative would be to use param = object@chain[1, j]
which is the start value for the parameters throughout the
initialization phase.